first part of explicit downcast & check

+++ src/catalog/catalog_entry/table_catalog_entry.cpp
@@ -183,13 +183,13 @@ vector<TypeId> TableCatalogEntry::GetTypes(const vector<column_t> &column_ids) {
 void TableCatalogEntry::Serialize(Serializer &serializer) {
 	serializer.WriteString(schema->name);
 	serializer.WriteString(name);
-	serializer.Write<uint32_t>(columns.size());
+	serializer.Write<uint32_t>((uint32_t)columns.size());
 	for (auto &column : columns) {
 		serializer.WriteString(column.name);
 		column.type.Serialize(serializer);
 		serializer.WriteOptional(column.default_value);
 	}
-	serializer.Write<uint32_t>(constraints.size());
+	serializer.Write<uint32_t>((uint32_t)constraints.size());
 	for (auto &constraint : constraints) {
 		constraint->Serialize(serializer);
 	}
